fluxbox: Remove
This commit remove fluxbox from my config options.

After using i3 and openbox for some time I decided to give up on fluxbox
due to it's been hard to setup, while i3 and openbox are easier/faster
to do so.

This commit also move gtk+ settings to the root dir.